requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
Werkzeug==3.1.3
google-ai-generativelanguage==0.6.15
google-genai==1.10.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-api-core==2.24.2
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
pillow==11.2.1
python-dotenv==1.1.0